requirements.txt
requests==2.25.1
python-decouple==3.3
psycopg2-binary==2.8.6
google-api-python-client
google-auth-httplib2
google-auth-oauthlib